# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Documentacion API Paises",
      "description": "Documentación realizada con swagger-ui-express: 4.6.3",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "lhezver@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "0.0.1"
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "tags": [
      {
        "name": "pais",
        "description": "Everything about your Pais",
        "externalDocs": {
          "description": "Find out more",
          "url": "http://swagger.io"
        }
      }
    ],
    "paths": {
      "/api/paises": {
        "get": {
          "tags": [
            "pais"
          ],
          "summary": "Returns all paises",
          "description": "Returns a map of status codes to quantities",
          "operationId": "getPaises",
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pais"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Pais not found"
            }
          }
        },
        "post": {
          "tags": [
            "pais"
          ],
          "summary": "Add a new pais",
          "description": "Add a new pais",
          "operationId": "addPais",
          "requestBody": {
            "description": "Create a new pais",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaisSinId"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pais"
                  }
                }
              }
            },
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/api/paises/{paisId}": {
        "get": {
          "tags": [
            "pais"
          ],
          "summary": "Find pais by ID",
          "description": "Returns a single pais",
          "operationId": "getPaisById",
          "parameters": [
            {
              "name": "paisId",
              "in": "path",
              "description": "ID of pais to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pais"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Pais not found"
            }
          }
        },
        "put": {
          "tags": [
            "pais"
          ],
          "summary": "Update a pais",
          "description": "Update a pais",
          "operationId": "putPais",
          "parameters": [
            {
              "name": "paisId",
              "in": "path",
              "description": "ID of pais to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "description": "Update a pais",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaisSinId"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pais"
                  }
                }
              }
            },
            "405": {
              "description": "Invalid input"
            }
          }
        },
        "delete": {
          "tags": [
            "pais"
          ],
          "summary": "Delete a pais",
          "description": "delete a pais",
          "operationId": "deletePais",
          "parameters": [
            {
              "name": "paisId",
              "in": "path",
              "description": "Pais id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Order not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Pais": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64",
              "example": 10
            },
            "nombre": {
              "type": "string",
              "example": "pais"
            },
            "habitantes": {
              "type": "integer",
              "format": "int64",
              "example": 10
            }
          }
        },
        "PaisSinId": {
          "type": "object",
          "properties": {
            "nombre": {
              "type": "string",
              "example": "pais"
            },
            "habitantes": {
              "type": "integer",
              "format": "int64",
              "example": 10
            }
          }
        }
      }
    }
  }